requirements.txt
torch==2.8.0
torchvision==0.23.0
transformers==4.30.2
Pillow==10.1.0
opencv-python-headless==4.8.1.78
sentence-transformers==2.2.2
chromadb-client==0.4.24.dev0
neo4j==5.12.0
flask==2.2.2
Werkzeug==2.3.7
waitress==2.1.2
streamlit==1.35.0
streamlit-autorefresh==1.0.1
pandas==2.2.0
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.0
huggingface_hub==0.15.1
